webhook-to-firestore-pipeline/requirements.txt
functions-framework==3.5.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pytz==2023.3.post1
google-cloud-firestore==2.14.0
Flask==3.0.1